fix bug
The restrictions that ISO C places on the second parameter to the
va_start() macro in header <stdarg.h> are different in this
International Standard. The parameter  parmN is the identifier of the
rightmost parameter in the variable parameter list of the function
definition (the one just before the  ...). If the parameter parmN is
declared with a function, array, or reference type, or with a type that
is not compatible with the type that results when passing an argument
for which there is no parameter, the behavior is undefined.